Use Go modules

+++ ptime.go
@@ -1,6 +1,6 @@
 // In the name of Allah
 
-// Persian Calendar v0.3
+// Persian Calendar v0.3.1
 // Please visit https://github.com/yaa110/go-persian-calendar for more information.
 //
 // Copyright (c) 2016 Navid Fathollahzade
@@ -635,7 +635,7 @@ func (t Time) MonthWeek() int {
 
 // YearWeek returns the week of year of t.
 func (t Time) YearWeek() int {
-	return int(math.Ceil(float64(float64(t.YearDay())+float64(t.FirstYearDay().Weekday())) / 7.0))
+	return int(math.Ceil(float64(t.YearDay()+int(t.FirstYearDay().Weekday())) / 7.0))
 }
 
 // RYearWeek returns the number of remaining weeks of the year of t.
